Extract some more methods.
